--- ___
Goal of the project is to generate a fractal structure using Python with some color added.
The hierarchical structure was reshaped to calculate an optimal random fractal pattern. Random fractals are method of choice for the purpose of computer image synthesis. The density of structure can be reduced and increased based on the given parameters.

You will like these project!

------